First pass at unified_mode support
Move all libraries into the resources folder
Move libaries back into the libraries folder
Move partials into the new _partials folder
Set unified_mode everywhere
Bump minimum supported version to 15.3
