+++ resources/docker_service_manager_execute.rb
