+++ resources/docker_service_manager_systemd.rb
